+++ src/main/java/kr/hhplus/be/server/application/order/OrderFacade.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.order.facade;
+package kr.hhplus.be.server.application.order;
 
 
 import kr.hhplus.be.server.coupon.dto.CouponDiscountInfo;
