+++ src/main/java/kr/hhplus/be/server/application/order/OrderResult.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.order.facade;
+package kr.hhplus.be.server.application.order;
 
 
 import kr.hhplus.be.server.order.dto.OrderInfo;
