+++ src/main/java/kr/hhplus/be/server/order/controller/OrderController.java
@@ -4,8 +4,8 @@ package kr.hhplus.be.server.order.controller;
 import kr.hhplus.be.server.common.response.ApiResponse;
 import kr.hhplus.be.server.common.response.ApiResponseCode;
 import kr.hhplus.be.server.order.controller.swagger.OrderSwaggerDocs;
-import kr.hhplus.be.server.order.facade.OrderFacade;
-import kr.hhplus.be.server.order.facade.OrderResult;
+import kr.hhplus.be.server.application.order.OrderFacade;
+import kr.hhplus.be.server.application.order.OrderResult;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
